extract IntUiBridgeChip
